removed separate file for filter
