+++ src/components/BlogList/BlogList.jsx
@@ -1,16 +1,16 @@
 import PropTypes from "prop-types";
 import { useDispatch, useSelector } from "react-redux";
 import useSearch from "hooks/useSearch";
+import { tagRemoved, tagselected } from "features/search/searchSlice";
 import BlogCard from "components/BlogCard/BlogCard";
 import Button from "components/Button/Button";
 import "./blogList.scss";
 import { tags } from "components/SelectBox/SelectBox";
-import { tagRemoved, tagselected } from "features/filter/filterSlice";
 
 const BlogList = ({ blogs }) => {
   const { query } = useSearch();
   const dispatch = useDispatch();
-  const { filteredTags } = useSelector((state) => state.filter);
+  const { filteredTags } = useSelector((state) => state.search);
 
   const handleSelect = (tag) => {
     const isSelected = filteredTags.includes(tag);
