+++ src/features/search/searchSlice.jsx
@@ -2,6 +2,7 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const INITIAL_STATE = {
   query: "",
+  filteredTags: [],
 };
 
 const searchSlice = createSlice({
@@ -11,8 +12,22 @@ const searchSlice = createSlice({
     updateQuery(state, action) {
       state.query = action.payload;
     },
+
+    tagselected: (state, action) => {
+      if (state.filteredTags.indexOf(action.payload) === -1) {
+        state.filteredTags.push(action.payload);
+      }
+    },
+
+    tagRemoved: (state, action) => {
+      const indexToRemove = state.filteredTags.indexOf(action.payload);
+
+      if (indexToRemove !== -1) {
+        state.filteredTags.splice(indexToRemove, 1);
+      }
+    },
   },
 });
 
 export default searchSlice.reducer;
-export const { updateQuery } = searchSlice.actions;
+export const { updateQuery, tagselected, tagRemoved } = searchSlice.actions;
