+++ src/store.js
@@ -3,7 +3,6 @@ import registerSlice from "features/register/registerSlice";
 import authSlice from "features/auth/authSlice";
 import blogsSlice from "features/blogs/blogsSlice";
 import searchSlice from "features/search/searchSlice";
-import filterSlice from "features/filter/filterSlice";
 
 const store = configureStore({
   reducer: {
@@ -11,7 +10,6 @@ const store = configureStore({
     auth: authSlice,
     blogs: blogsSlice,
     search: searchSlice,
-    filter: filterSlice,
   },
 });
